Add some basic test cases
This commit adds some basic test cases when building the Dockerfile.
Since rabbitmqadmin depends on python, we temporarily pull it in
during the build. We also have rabbitmqadmin in our repo instead of
depending on a version in the RabbitMQ repo since we need to, until we properly
fix it, disable certificate verification when performing the tests since
python versions > 2.7.9 have enabled it by default.

I also shuffled things around in the Dockerfile to make it more
cache-friendly.

See https://www.python.org/dev/peps/pep-0476