Don't sleep for synchronization, use exec
This adds actual synchronization instead of using `sleep` to wait until
things are up.

The main benefit of doing so is that the tests run faster (rather than
wait 3 * 25 seconds and therefore run in > 75 seconds / test, each test
runs in about 7 seconds in my testing)

Of course, the downside is that we need to be more accurate in ensuring
that things time and exit properly, etc. So, this makes a few other
changes:

- It waits until RabbitMQ is loaded in `rabbitmqctl status` before
  adding users, queues, etc.
- It waits until RabbitMQ exits after initilizing it.
- It transfers signals (by using `exec`) to `rabbitmq-server` when
  running RabbitMQ (incidentally, this should also make things work
  better with Captain Comeback, although I'm still testing this).

It also:

- Ensures we actually use the wrapper to run RabbitMQ when running
  tests (after all; it's an important part of the image).
- Ensures we actually use `RABBITMQ_MNESIA_BASE` in all tests.
- Ensures epmd exits after each test. This isn't strictly speaking
  required, but it ensures we have a cleaner slate when running each
  test, which better reflects the environment RabbitMQ will be running
  in.