Prefer certificates passed in the environment
We're approaching the renewal deadline for the *.aptible.in certificate
used by (among others) RabbitMQ databases, so we'll have to be able to
restart them with a different certificate.

This patch ensures that when RabbitMQ restarts, it uses the certificate
found in the environment if there is one (and otherwise falls back to
using the certificate from the filesystem). Until now, we were doing
the opposite: a certificate from the filesystem would be preferred
over a certificate from the environment.